# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects": {
      "post": {
        "operationId": "ProjectsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new project",
        "tags": [
          "Projects"
        ]
      },
      "get": {
        "operationId": "ProjectsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of projects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProjectResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get all projects",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "summary": "Get a single project by ID",
        "tags": [
          "Projects"
        ]
      },
      "put": {
        "operationId": "ProjectsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not your project"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}/applications": {
      "post": {
        "operationId": "ProjectsController_apply",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Application submitted."
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Apply for a project",
        "tags": [
          "Projects"
        ]
      },
      "get": {
        "operationId": "ProjectsController_getApplications",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of applications."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not your project"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get applications for a project (Owner only)",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}/applications/{appId}": {
      "put": {
        "operationId": "ProjectsController_processApplication",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Application processed."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not your project"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Process an application (Owner only)",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}/members": {
      "post": {
        "operationId": "ProjectsController_assignRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role assigned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMemberResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Only PM or Owner can assign roles"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign a role to a project member",
        "tags": [
          "Projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Task Bounty Platform API",
    "description": "The Nền tảng Quản lý công việc và Giao việc tự do API description.\n\n### Real-time Events (WebSocket)\nThis platform provides real-time updates via Socket.io.\n- **Endpoint**: `ws://<host>:<port>/notifications`\n- **Transports**: `['websocket', 'polling']`\n\n#### Events to Listen For\n1. **`risk-alert`**: Emitted when the AI detects a high risk score for a task.\n   - Payload: `{ pmId: string, message: string }`\n2. **`balance-warning`**: Emitted when the PM's escrow balance drops below a required threshold.\n   - Payload: `{ pmId: string, message: string }`\n",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "lastPage": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "lastPage"
        ]
      },
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "DeFi Dashboard"
          },
          "description": {
            "type": "string",
            "example": "A dashboard to track DeFi metrics."
          },
          "budget": {
            "type": "number",
            "example": 1000
          },
          "type": {
            "type": "string",
            "example": "PUBLIC",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          }
        },
        "required": [
          "title",
          "description",
          "budget",
          "type"
        ]
      },
      "ProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-1234"
          },
          "pmId": {
            "type": "string",
            "example": "uuid-pm-1234"
          },
          "title": {
            "type": "string",
            "example": "DeFi Dashboard"
          },
          "description": {
            "type": "string",
            "example": "A dashboard to track DeFi metrics."
          },
          "budget": {
            "type": "number",
            "example": 1000
          },
          "type": {
            "type": "string",
            "example": "PUBLIC"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-07-30T10:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-07-30T10:00:00Z"
          }
        },
        "required": [
          "id",
          "pmId",
          "title",
          "description",
          "budget",
          "type",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedProjectResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "AssignRoleDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "uuid-user-1234"
          },
          "role": {
            "type": "string",
            "example": "LEAD_DEV",
            "enum": [
              "PM",
              "LEAD_DEV",
              "REVIEWER",
              "DEV"
            ]
          }
        },
        "required": [
          "userId",
          "role"
        ]
      },
      "ProjectMemberResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-member-1234"
          },
          "projectId": {
            "type": "string",
            "example": "uuid-project-1234"
          },
          "userId": {
            "type": "string",
            "example": "uuid-user-1234"
          },
          "role": {
            "type": "string",
            "example": "LEAD_DEV"
          }
        },
        "required": [
          "id",
          "projectId",
          "userId",
          "role"
        ]
      },
...
    }
  }
}